# fix(spanner): row mismatch in SelectAll using custom type