--- a/spanner/row_test.go
+++ b/spanner/row_test.go
@@ -2241,6 +2241,25 @@ func BenchmarkColumn(b *testing.B) {
 	}
 }
 
+type CustomType[T any] struct {
+	Val T
+}
+
+func (n *CustomType[T]) DecodeSpanner(input any) error {
+	switch val := input.(type) {
+	case T:
+		n.Val = val
+		return nil
+	case *T:
+		if val == nil {
+			return nil
+		}
+		n.Val = *val
+		return nil
+	}
+	panic("n/a")
+}
+
 func TestSelectAll(t *testing.T) {
 	skipUnsupportedPGTest(t)
 	type args struct {
@@ -2255,6 +2274,13 @@ func TestSelectAll(t *testing.T) {
 		Col3 string
 		Col4 time.Time
 	}
+
+	type testStructWithCustom struct {
+		Col1 int64
+		COL2 float64
+		Col3 CustomType[string]
+	}
+
 	type testStructWithTag struct {
 		Col1 int64     `spanner:"tag1"`
 		Col2 float64   `spanner:"Tag2"`
@@ -2422,6 +2448,44 @@ func TestSelectAll(t *testing.T) {
 				{Col1: 3, COL2: 3.3, Col3: "value3"},
 				{Col1: 1, COL2: 1.1, Col3: "value"},
 				{Col1: 2, COL2: 2.2, Col3: "value2"},
+			},
+		},
+		{
+			name: "success: using destination with custom type with custom decoder with some null columns",
+			args: args{
+				destination: &[]*testStructWithCustom{},
+				mock: newMockIterator(
+					&Row{
+						[]*sppb.StructType_Field{
+							{Name: "Col1", Type: intType()},
+							{Name: "Col2", Type: floatType()},
+							{Name: "Col3", Type: stringType()},
+						},
+						[]*proto3.Value{intProto(3), floatProto(3.3), stringProto("value3")},
+					},
+					&Row{
+						[]*sppb.StructType_Field{
+							{Name: "Col1", Type: intType()},
+							{Name: "Col2", Type: floatType()},
+							{Name: "Col3", Type: stringType()},
+						},
+						[]*proto3.Value{intProto(1), floatProto(1.1), nullProto()},
+					},
+					&Row{
+						[]*sppb.StructType_Field{
+							{Name: "Col1", Type: intType()},
+							{Name: "Col2", Type: floatType()},
+							{Name: "Col3", Type: stringType()},
+						},
+						[]*proto3.Value{intProto(2), floatProto(2.2), stringProto("value2")},
+					},
+					iterator.Done,
+				),
+			},
+			want: &[]*testStructWithCustom{
+				{Col1: 3, COL2: 3.3, Col3: CustomType[string]{"value3"}},
+				{Col1: 1, COL2: 1.1, Col3: CustomType[string]{}},
+				{Col1: 2, COL2: 2.2, Col3: CustomType[string]{"value2"}},
 			},
 		},
 		{