@sakthivelmanii 2025-05-08
@rahul2393 @olavloite I changed the approach and set the value using reflection instead of re-initializing memory.

Benchmark results

```
goos: darwin
goarch: arm64
pkg: cloud.google.com/go/spanner
cpu: Apple M3 Pro
                             │ 10_runs_bench_using_select_all_main.txt │ 10_runs_bench_using_select_all.txt  │
                             │                 sec/op                  │   sec/op     vs base                │
Scan100RowsUsingSelectAll-12                               11.72µ ± 1%   14.78µ ± 0%  +26.06% (p=0.000 n=20)

                             │ 10_runs_bench_using_select_all_main.txt │ 10_runs_bench_using_select_all.txt │
                             │                  B/op                   │      B/op       vs base            │
Scan100RowsUsingSelectAll-12                              12.30Ki ± 0%     12.30Ki ± 0%  ~ (p=1.000 n=20) ¹
¹ all samples are equal

                             │ 10_runs_bench_using_select_all_main.txt │ 10_runs_bench_using_select_all.txt │
                             │                allocs/op                │   allocs/op     vs base            │
Scan100RowsUsingSelectAll-12                                222.0 ± 0%       222.0 ± 0%  ~ (p=1.000 n=20) ¹
¹ all samples are equal
```